blazor working

+++ FreightTransport_BLL/DTOs/RouteDTO.cs
@@ -4,7 +4,7 @@ namespace FreightTransport_BLL.DTOs
     public class RouteDTO
     {
         public int Id { get; set; }
-        public int Distance { get; set; }
+        public double Distance { get; set; }
         public int Cost { get; set; }
 
         public int DestinationCityId { get; set; }
